refactor(discord): add file extension on import paths
