+++ src/discord/DiscordBot.ts
@@ -5,15 +5,17 @@
  * 
  */
 
-import log4js from 'log4js';
-import { Client, Permissions, Guild, GuildChannel, ThreadChannel, CommandInteraction, ApplicationCommandData, ApplicationCommandPermissionData, CreateRoleOptions, MessageEmbed, MessageActionRow, MessageButton, DiscordAPIError, Message, Role, TextChannel, GuildMember, ButtonInteraction } from 'discord.js';
-import config from 'config';
-
-import { IIrcClient, LobbyStatus, Player } from '..';
-import { OahrDiscord } from './OahrDiscord';
-import { setContext } from './DiscordAppender';
-import { BotCommands } from './BotCommand';
-import { BanchoResponse, BanchoResponseType } from '../parsers';
+ import log4js from 'log4js';
+ import { Client, Permissions, Guild, GuildChannel, ThreadChannel, CommandInteraction, ApplicationCommandData, ApplicationCommandPermissionData, CreateRoleOptions, MessageEmbed, MessageActionRow, MessageButton, DiscordAPIError, Message, Role, TextChannel, GuildMember, ButtonInteraction } from 'discord.js';
+ import config from 'config';
+ 
+ import { IIrcClient } from '../IIrcClient.js';
+ import { LobbyStatus } from '../Lobby.js';
+ import { Player } from '../Player.js';
+ import { OahrDiscord } from './OahrDiscord.js';
+ import { setContext } from './DiscordAppender.js';
+ import { BotCommands } from './BotCommand.js';
+ import { BanchoResponse, BanchoResponseType } from '../parsers/CommandParser.js';
 
 const logger = log4js.getLogger("discord");
