+++ src/discord/OahrDiscord.ts
@@ -1,7 +1,9 @@
-import { IIrcClient, LobbyStatus, Player } from '..';
+import { IIrcClient } from '../IIrcClient.js';
+import { LobbyStatus } from '../Lobby.js';
+import { Player } from '../Player.js';
 import log4js from 'log4js';
-import { OahrBase } from '../cli/OahrBase';
-import { OahrSharedObjects } from './DiscordBot';
+import { OahrBase } from '../cli/OahrBase.js';
+import { OahrSharedObjects } from './DiscordBot.js';
 import { MessageActionRow, MessageButton, MessageEmbed } from 'discord.js';
 import { MessageButtonStyles } from 'discord.js/typings/enums';
